refactor: extract update handling to separate file
