+++ bitrise/print.go
@@ -2,7 +2,6 @@ package bitrise
 
 import (
 	"fmt"
-	"regexp"
 	"strings"
 	"time"
 	"unicode/utf8"
@@ -12,9 +11,7 @@ import (
 	"github.com/bitrise-io/go-utils/colorstring"
 	log "github.com/bitrise-io/go-utils/log"
 	"github.com/bitrise-io/go-utils/stringutil"
-	"github.com/bitrise-io/go-utils/versions"
 	stepmanModels "github.com/bitrise-io/stepman/models"
-	ver "github.com/hashicorp/go-version"
 )
 
 const (
@@ -25,35 +22,6 @@ const (
 //------------------------------
 // Util methods
 //------------------------------
-
-func isUpdateAvailable(stepInfo stepmanModels.StepInfoModel) bool {
-	if stepInfo.LatestVersion == "" {
-		return false
-	}
-
-	if stepInfo.Version != stepInfo.EvaluatedVersion {
-		re := regexp.MustCompile(`\d+`)
-		components := re.FindAllString(stepInfo.Version, -1)
-		normalized := strings.Join(components, ".")
-		locked, _ := ver.NewSemver(normalized)
-		latest, _ := ver.NewSemver(stepInfo.LatestVersion)
-
-		switch len(components) {
-		case 1:
-			return locked.Segments()[0] < latest.Segments()[0]
-		case 2:
-			return locked.Segments()[0] < latest.Segments()[0] || locked.Segments()[1] < latest.Segments()[1]
-
-		}
-	}
-
-	res, err := versions.CompareVersions(stepInfo.Version, stepInfo.LatestVersion)
-	if err != nil {
-		log.Errorf("Failed to compare versions, err: %s", err)
-	}
-
-	return (res == 1)
-}
 
 func getTrimmedStepName(stepRunResult models.StepRunResultsModel) string {
 	iconBoxWidth := len("   ")
